+++ apps/web/src/services/cost-explorer/components/AdvancedSettingsAdjustmentGroupForm.vue
@@ -19,8 +19,8 @@ import type {
 } from '@/lib/variable-models/variable-model-menu-handler';
 
 import AdvancedSettingsAdjustmentsForm from '@/services/cost-explorer/components/AdvancedSettingsAdjustmentsForm.vue';
-import { useCostReportConfigQuery } from '@/services/cost-explorer/composables/queries/use-cost-report-config-query';
-import { useReportAdjustmentQuery } from '@/services/cost-explorer/composables/queries/use-report-adjustment-query';
+import { useCostReportConfigQuery } from '@/services/cost-explorer/composables/use-cost-report-config-query';
+import { useReportAdjustmentQuery } from '@/services/cost-explorer/composables/use-report-adjustment-query';
 import { useAdvancedSettingsPageStore } from '@/services/cost-explorer/stores/advanced-settings-page-store';
